font-patcher: Be explicit what shall be transformed
[why]
In the patch loop we use two methods to do something with the glyphs
- the glyph object directly (i.e. font[codepoint])
- a selection of glyphs in the font

Using the selection is a bit anonymous and depends on lines far away
(where the selection is set); direct glyph access seems to be easier
understood and is used almost everywhere.

Direct glyph access can not be used for copy() and paste(), but for
everything else.

[how]
The direct glyph addressing is already in use almost for everything, but
not for the transform() calls. This is changed.
Now we transform a specific glyph and not 'all selected glyphs' (with
selection = that specific glyph).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>