font-patcher: Warn if sourcefont is a VF
[why]
fontforge is not really able to work with OpenType variable fonts, at
least not with all. Some support is available as MM, for older formats.
But anyhow we do not really create a patched variable font but a fixed
font.

People might ignore all the errors Fontforge throws on opening, so an
explicit message might be in order.

[how]
It is not possible to detect a VF input font with current fontforge
reliably. Instead we search for the existence of one of the tables that
are  needed for a variable font. We can not rely on STAT, because that
might be also used in fixed fonts.

Some fontforges might crash on VFs, so we give a warning before we even
open the font and one after the patched font has been created.

Fixes: #512

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>