font-patcher: Fix checksum algo
[why]
For some fonts the calculated checksum(s) was wrong.

[how]
Misplaced the multiplier ... the LSB shall not be shifted of course, but
the current code shifts as last action: Also the LSB is shifted.

First shift (old) accumulated content, then add new data.
In that way the last added data (LSB) is not shifted.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>