archive-fonts: Fix missing licence files
[why]
We have 'license' is the archives, but the 'licence' files are missing.
We allow all-caps and all-lowercase, but not camelcase 'License'.
Sometimes we have OFL.txt which is missing.

[how]
Enable both spellings and also camelcase.
Enable OTF text files.

Also fix misunderstanding of regex vs glob ([] vs {}), where the glob
needs a comma and the regex does not.

Add missing LICENSE files for 2 fonts.

Fixes: #979

Reported-by: Jeremy Thorn <moralground>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>