CI: No not run the release workflow on forks
[why]
When you have a fork of this repository and push anything to master the
release workflow is automatically run. This is usually not what people
want. They want to push some changes to the fork's master and create a
Pull Request.

[how]
Only run the release workflow on the original repo.

If they really want to create their own release they would have to
remove this line (or adapt).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>