font-patcher: Fix GlyphToScale for 'xy' scales
[why]
The scale-glyph-data is used only for 'pa' scales, but thereafter used
for all shifts, even if the scaling has been 'x' or 'y' or both.

As we do not use GlyphToScale for anything but 'pa' scaled glyphs that
should not make any difference right now. But it will be an obscure bug
if we ever want to handle the Powerline symbols with a scale group.

I do not know if that will ever happen, but I tried it whilst
experimenting spending hours on finding this bug.

[how]
Access the GlyphToScale data and use it even for 'x' and 'y' scaling, if
we have it for the particular glyph.

[note]
Also change 'invalid' flag from False to None.
Also use 'is None' or 'is not None' for comparisons with None.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>