CI: Deactivate release workflow
[why]
More often than not (i.e. almost always) I cancel the release workflow.

The problem is that a release workflow will patch all fonts, and all
fonts will differ (at least in the timestamp) and so the repo will grow
A LOT. Usually you want to be really conciously deciding that the growth
is really warranted, and no automatic can do that.

I guess one could trigger rebuilding the zip archives but not commiting
the newly patched fonts back to the repo, but that is also a strange
situation.

[how]
Release has now only a workflow-dispatch trigger, that must be clicked
if a release workflow shall run.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>