name-parser: Shorten 'Oblique' to 'Obl'
[why]
In a lot fonts when Oblique is moved to the base part of the name the
family gets too long. Since the base part is never shortened it will
stay there and make some fonts will end up with too ong bas families:

DejaVuSansMonoOblique
IosevkaTermOblique
OpenDyslexicOblique
ProggyCleanOblique

[how]
If the name is shortened, take the short form of oblique and add that to
the base name.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>