Fix standardize-and-complete-readmes
[why]
A lot of fonts have a wrong RFN statement and name substitude in their
created info file, for example IBM3270:

```
--- a/patched-fonts/3270/Condensed/font-info.md
+++ b/patched-fonts/3270/Condensed/font-info.md
@@ -1,6 +1,20 @@
 # 3270

+## Why `Anonymice Pro` and not `Anonymous Pro`?
+
+What's in a name? The reason for the name change is to comply with the SIL Ope
```

[how]
The rfn information is stored in the config.cfg files - i.e. variables.
They are reset if the next font does not have a config.cfg file; but
they are not reset if the next font DOES have a config.cfg file but does
not set anything.

Reset the values before reading the config.cfg file.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>